fix: centralise STATUS_COLORS/VALID_STATUSES, fix TTL race, fix inspector position lag
- Move STATUS_COLORS from store.js to constants.js (re-exported for compatibility)
- Import VALID_ROLES/VALID_STATUSES from constants in inferStatus.js and normalizePost.js,
  eliminating duplicate definitions across three files
- Fix TTL expiry race: re-check fresh expiresAt before clearing to prevent a renewed
  working status from being wiped by a stale interval snapshot
- Fix AgentInspector anchor position: use agent.targetPosition while isMoving so the
  popover and connection line follow the agent's destination, not the departure point
