+++ src/systems/constants.js
@@ -25,3 +25,10 @@ export const TIME_CHECK_INTERVAL = 60000               // 1 min
 export const STATUS_POLL_INTERVAL = 2000   // ms
 export const VALID_ROLES = ['pm', 'arch', 'dev', 'qa', 'ops', 'res', 'gate']
 export const VALID_STATUSES = ['idle', 'working', 'blocked', 'done']
+
+export const STATUS_COLORS = {
+  idle: '#888',
+  working: '#EF9F27',
+  done: '#5CB88A',
+  blocked: '#E24B4A',
+}
